requirements.txt
instagrapi==2.1.2
python-dotenv==1.0.1
pydantic==2.7.1
click==8.1.7